requirements.txt
python-dotenv
discord-py-interactions
aiohttp
arrow
